+++ Runtime/SpriteShape.cs
@@ -130,7 +130,10 @@ namespace UnityEngine.U2D
                 {
                     Sprite sprite = sprites[i];
                     if (sprite)
+                    {
+                        hashCode ^= (i + 1);
                         hashCode ^= sprite.GetHashCode();
+                    }
                 }
             }
 
@@ -196,7 +199,10 @@ namespace UnityEngine.U2D
                 {
                     Sprite sprite = sprites[i];
                     if (sprite)
+                    {
+                        hashCode ^= (i + 1);
                         hashCode ^= sprite.GetHashCode();
+                    }
                 }
             }